Fix cannot add room

+++ backend/src/main/java/com/example/exam_online/controller/UserController.java
@@ -34,7 +34,7 @@ public class UserController {
     private JwtTokenProvider tokenProvider;
 
     @GetMapping("/findUser/{userId}")
-    public ResponseHandler<UserDto> findUserById(@PathVariable int userId) throws CustomException {
+    public ResponseHandler<UserDto> findUserById(@PathVariable Long userId) throws CustomException {
         User user = userService.findById(userId);
         UserDto userDto = mapper.map(user, UserDto.class);
         ResponseHandler<UserDto> responseHandler = new ResponseHandler<UserDto>("successfully found user",
